+++ ui/mods/com.pa.quitch.galactic-war-multiplanetary-systems/shared/build_map_list.js
@@ -19,7 +19,7 @@ if (!multiplanetarySystemTabsLoaded) {
         coherentFilePath
       ) {
         if (numberOfPlanets > 1) {
-          multiplanetaryMaps.push(coherentFilePath);
+          return coherentFilePath;
         }
       };
 
@@ -30,11 +30,61 @@ if (!multiplanetarySystemTabsLoaded) {
             startingPlanets += 1;
           }
           if (startingPlanets > 1) {
-            multiStartMaps.push(coherentFilePath);
-            break;
+            return coherentFilePath;
           }
         }
       };
+
+      var defaultMultiplanetary = [];
+      var defaultMultiStart = [];
+
+      var processDefaultSystems = function (systems) {
+        _.forEach(systems, function (system) {
+          var multiPlanetResult = checkForMultiplePlanets(
+            system.planets.length,
+            system
+          );
+          var multiStartResult = checkForMultiplanetarySpawns(
+            system.planets,
+            system
+          );
+          if (_.isObject(multiPlanetResult)) {
+            defaultMultiplanetary.push(multiPlanetResult);
+          }
+          if (_.isObject(multiStartResult)) {
+            defaultMultiStart.push(multiStartResult);
+          }
+        });
+      };
+
+      // Scan My Systems and PA for maps when populated
+      model.userSystems.subscribe(function (systems) {
+        processDefaultSystems(systems);
+      });
+      model.premadeSystems.subscribe(function (systems) {
+        processDefaultSystems(systems);
+      });
+
+      // Add My Systems and PA maps when tabs are ready
+      var addedDefaultMultiSystems = false;
+      var addedDefaultMultiStarts = false;
+      model.cShareSystems_tabsIndex.subscribe(function (tabs) {
+        if (
+          addedDefaultMultiSystems === false ||
+          addedDefaultMultiStarts === false
+        ) {
+          _.forEach(tabs, function (tab) {
+            if (tab.name === mapTabOne && addedDefaultMultiSystems === false) {
+              tab.systems(tab.systems().concat(defaultMultiplanetary));
+              addedDefaultMultiSystems = true;
+            }
+            if (tab.name === mapTabTwo && addedDefaultMultiStarts === false) {
+              tab.systems(tab.systems().concat(defaultMultiStart));
+              addedDefaultMultiStarts = true;
+            }
+          });
+        }
+      });
 
       api.file.list("/ui/mods/", true).then(function (fileList) {
         var deferredQueue = [];